fix weird and remove seperate sdl3

+++ project/include/ui/GamepadEvent.h
@@ -28,7 +28,7 @@ namespace lime {
 		int id;
 		GamepadEventType type;
 		double axisValue;
-		double timestamp;
+		uint64_t timestamp;
 
 		static ValuePointer* callback;
 		static ValuePointer* eventObject;
